Create CH-120 payments Durations
In challenge 60, we attempted to calculate the source of payment for each receipt. In this challenge, we are going to calculate the average duration of payment for each receipt. The customer receipt costs and their payments are provided in tables 1 and 2. We will match the payment to the receipt based on the date and calculate the payment duration for each receipt. 
For example, all of receipt C1 is paid by Payment P1, 47 days before the receipt. For receipt C5, out of the total $374, $337 is paid by P4 (49 days before) and $37 is paid by P5 (27 days before), resulting in an average payment duration of -47.8 days.