Create CH-139 Custom Grouping
The stock prices for the given dates are provided in the table. Extract the highest increase and decrease in stock price either in a single day or over consecutive days with the same upward or downward trend.
 
A trend is defined as a movement in the same direction for more than two consecutive dates.
In this example, the highest decrease occurred between 3/1 and 4/1, while the highest increase happened between 4/1 and 8/1, both following a consistent upward trend.