Create CH-290 Flight Planning
The Question table contains a flight schedule for a single day from multiple locations. Each location is in a different time zone, provided in the Table 2. The departure times in the From column are given in the local time of the departure city.
Find all possible routes from City A to City B, taking into account the time zone differences to ensure correct sequencing of flights.